Refactor + Simplify module import for library consumption
