+++ examples/zenoh-ts/package.json
@@ -14,6 +14,7 @@
     "vite": "^5.2.12"
   },
   "dependencies": {
-    "tslog": "^4.9.3"
+    "tslog": "^4.9.3",
+    "zenoh": "file:../../esm/"
   }
 }
